+++ 4_Delegates/4_Delegates.csproj
@@ -7,10 +7,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="IteaMessanger\Group.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\1_Collections\1_Collections.csproj" />
   </ItemGroup>
